django/requirements.txt
amply==0.1.2
amqp<6.0.0,>=5.0.6
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
celery==5.0.5
Django==2.2
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.6.2
django-admin-tools==0.8.1
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-grappelli==2.13.4
django-tagulous==1.0.0
djangorestframework==3.12.4
docutils==0.16
draftfast==3.5.7
flower==1.0.0
idna==2.6
isort==4.3.21
joblib==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
psycopg2-binary==2.7.4
PuLP==2.2
pydfs-lineup-optimizer==3.3.0
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.5.3
pytz==2020.1
redis==3.2.1
requests==2.18.4
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.22
uwsgi>=2.0.18
wrapt==1.12.1
